+++ src/ChessResources/ChessHistoryTracker/BoardStateChanges/BoardStateChange.java
@@ -3,18 +3,18 @@ package ChessResources.ChessHistoryTracker.BoardStateChanges;
 import ChessResources.Pieces.PieceData;
 
 public class BoardStateChange{
-    private int piece;
+    private short piece;
     private int spaceIdArriveAt;
     private int spaceId;
 
-    public BoardStateChange(int piece, int spaceId, int spaceIdArriveAt)
+    public BoardStateChange(short piece, int spaceId, int spaceIdArriveAt)
     {
         this.piece = piece; //should always be a copy.
         this.spaceId = spaceId;
         this.spaceIdArriveAt = spaceIdArriveAt;
     }
 
-    public int getPiece(){return piece;}
+    public short getPiece(){return piece;}
     public int getSpaceIdArriveAt(){return spaceIdArriveAt;}
     public int getSpaceId(){return spaceId;}
     @Override
