+++ src/ChessResources/PreCalc.java
@@ -22,10 +22,10 @@ public class PreCalc {
     public static int[][] KNIGHT_MOVES = new int[ChessBoard.BOARD_SIZE*ChessBoard.BOARD_SIZE][];
     public static int[][] KING_MOVES = new int[ChessBoard.BOARD_SIZE*ChessBoard.BOARD_SIZE][];
 
-    public static final int[] FEN_MAP = new int[128];
-    public static final int[] WHITE_THREAT_IDS = { PieceData.BPAWN, PieceData.BQUEEN, PieceData.BBISHOP,
+    public static final short[] FEN_MAP = new short[128];
+    public static final short[] WHITE_THREAT_IDS = { PieceData.BPAWN, PieceData.BQUEEN, PieceData.BBISHOP,
             PieceData.BROOK, PieceData.BKNIGHT, PieceData.BKING };
-    public static final int[] BLACK_THREAT_IDS = { PieceData.WPAWN, PieceData.WQUEEN, PieceData.WBISHOP,
+    public static final short[] BLACK_THREAT_IDS = { PieceData.WPAWN, PieceData.WQUEEN, PieceData.WBISHOP,
             PieceData.WROOK, PieceData.WKNIGHT, PieceData.WKING };
 
     public static final PieceData[] PIECE_ID_TO_PIECE_DATA_MAP = new PieceData[PieceData.MAX_PIECES];
